refactor: streamline webhook sending logic and improve error handling

+++ src/services/webhooks.py
@@ -12,6 +12,7 @@ from ..data.services.webhooks_db import (
 )
 from ..data.services.monitors_db import get_monitor_by_id as _get_monitor_by_id_
 from ..data.schemas import MonitorSchema
+from ..data.models import Webhook
 from ..utils.cache import CacheWebhookUser, CacheWebhookMonitor
 from ..exceptions import HTTPException
 
@@ -41,26 +42,34 @@ def send_webhooks(test: bool = False, **kwargs) -> None:
     """
     Envía los webhooks a las urls especificadas.
     """
-    def send_webhook_test() -> None:
-        with Session(engine) as session:               
-            webhook = get_webhook(session, kwargs.get('token_user'))
+
+    def send_data(data: list, webhook: Webhook) -> None:
+        try:
+            logger.info(f'Enviando webhook a: {webhook.url}')
+            asyncio.run(
+                send_webhook(webhook.url, webhook.token, webhook.certificate_ssl, {'monitors': data})
+            )
+        except Exception as e:
+            logger.error(f'Error al enviar el webhook: {str(e)}')
+            raise e
+
+    def send_webhook_test(token_user: str) -> None:
+        with Session(engine) as session:
+            webhook = get_webhook(session, token_user)
             if not webhook:
                 raise ValueError('Webhook no encontrado')
-            
+
             data = []
             for m in webhook.monitors:
                 monitor = _get_monitor_by_id_(session, m.monitor_id)
                 data.append(MonitorSchema(rounded_price=False).dump(monitor))
 
-            try:
-                asyncio.run(
-                    send_webhook(webhook.url, webhook.token, webhook.certificate_ssl, {'monitors': data})
-                )
-            except Exception as e:
-                raise e
+            send_data(data, webhook)
 
     if test:
-        send_webhook_test()
+        token_user = kwargs.get('token_user')
+        if token_user:
+            send_webhook_test(token_user)
         return
 
     with Session(engine) as session:
@@ -72,8 +81,8 @@ def send_webhooks(test: bool = False, **kwargs) -> None:
         for webhook in get_all_webhooks(session):
             if not webhook.status:
                 continue
+            
             data = []
-
             for m in webhook.monitors:
                 if m.monitor_id not in monitor_webhooks:
                     continue
@@ -85,19 +94,18 @@ def send_webhooks(test: bool = False, **kwargs) -> None:
                 monitor = _get_monitor_by_id_(session, m.monitor_id)
                 monitors_ids_save[m.monitor_id] = MonitorSchema(rounded_price=False).dump(monitor)
                 data.append(monitors_ids_save[m.monitor_id])
-            try:
-                asyncio.run(
-                    send_webhook(webhook.url, webhook.token, webhook.certificate_ssl, {'monitors': data})
-                )
-            except Exception as e:
-                logger.error(f'Error al enviar el webhook: {str(e)}')
 
+            if not data:
+                continue
+
+            try:
+                send_data(data, webhook)
+            except Exception as e:
                 cache_user = CacheWebhookUser(webhook.id)
                 if not cache_user.is_intents_webhook_limit():
                     cache_user.set()
                 else:
+                    logger.info(f'Webhook desactivado por superar el límite de intentos: {webhook.url}')
                     change_webhook_status(session, webhook.id, False)
                     cache_user.delete()
-                    logger.info(f'Webhook desactivado por superar el límite de intentos: {webhook.url}')
-
         CacheWebhookMonitor().delete_all_monitor_webhook()
